Fix 6 Kotlin compilation errors from batch 5
- AbstractPhoneContact: getLong instead of getInt for _ID (getLookupUri needs Long)
- Transferable: @JvmField requires all companion properties to use it; drop const
- MainThreadExecutor: execute() must return Unit, not Boolean (remove = shorthand)
- UriHelper: ImmutableMap disallows null values; use LinkedHashMap instead
- NickValidityChecker: room.local/domain are platform nullables, add !!
- StanzaIdManager: account.jid is platform nullable, add !!

+++ src/main/java/eu/siacs/conversations/android/AbstractPhoneContact.kt
@@ -12,7 +12,7 @@ abstract class AbstractPhoneContact protected constructor(cursor: Cursor) {
 
     init {
         @Suppress("DEPRECATION")
-        val phoneId = cursor.getInt(cursor.getColumnIndex(ContactsContract.Data._ID))
+        val phoneId = cursor.getLong(cursor.getColumnIndex(ContactsContract.Data._ID))
         @Suppress("DEPRECATION")
         val lookupKey = cursor.getString(cursor.getColumnIndex(ContactsContract.Data.LOOKUP_KEY))
         lookupUri = ContactsContract.Contacts.getLookupUri(phoneId, lookupKey)
